feat: add Room dependencies and data module to presentation
This commit introduces the Room persistence library by adding its dependencies to the `gradle/libs.versions.toml` file.

Additionally, the `data` module is now included as an implementation dependency in the `presentation/build.gradle.kts` file. This allows the presentation layer to access data layer components, such as repositories.

Finally, an import statement for `InMemoryRecentSearchRepository` was reordered in `SearchListViewModel.kt`.